Chnages made to cryptography library.
Not done still, have to upload more files.

+++ externals/xz-5.2.2/src/common/tuklib_mbstr_width.c
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-/// \file       tuklib_mbstr_width.c
+/// \file       tuklib_mstr_width.c
 /// \brief      Calculate width of a multibyte string
 //
 //  Author:     Lasse Collin
@@ -11,7 +11,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 #include "tuklib_mbstr.h"
-#include <string.h>
 
 #if defined(HAVE_MBRTOWC) && defined(HAVE_WCWIDTH)
 #	include <wchar.h>
@@ -51,7 +50,7 @@ tuklib_mbstr_width(const char *str, size_t *bytes)
 		if (wc_width < 0)
 			return (size_t)-1;
 
-		width += (size_t)wc_width;
+		width += wc_width;
 	}
 
 	// Require that the string ends in the initial shift state.
